chore: use one mod for rate limit
